# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "List of tasks",
                "parameters": [
                    {
                        "name": "filters[status_id]",
                        "in": "query",
                        "description": "filter by task status",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filters[priority_id]",
                        "in": "query",
                        "description": "filter by task priority",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filters[title]",
                        "in": "query",
                        "description": "filter by task title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[description]",
                        "in": "query",
                        "description": "filter by task description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "filter for sorting task by columns",
                        "required": false,
                        "schema": {
                            "type": "object"
                        },
                        "example": [
                            {
                                "priority_id": "asc"
                            }
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "status_id": {
                                                        "type": "integer"
                                                    },
                                                    "priority_id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object",
                                                "example": {
                                                    "id": 1,
                                                    "status_id": 1,
                                                    "priority_id": 3,
                                                    "title": "Some title",
                                                    "description": "Some description"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create task",
                "requestBody": {
                    "request": "some request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "parent_task_id": {
                                                "type": "integer"
                                            },
                                            "priority_id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "parent_task_id": null,
                                            "priority_id": 3,
                                            "title": "Some title",
                                            "description": "Some description"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "status_id": {
                                                    "type": "integer"
                                                },
                                                "priority_id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "status_id": 1,
                                                "priority_id": 3,
                                                "title": "Some title",
                                                "description": "Some description"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get task info",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "status_id": {
                                                    "type": "integer"
                                                },
                                                "priority_id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "status_id": 1,
                                                "priority_id": 3,
                                                "title": "Some title",
                                                "description": "Some description"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Task"
                ],
                "summary": "Update task info",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "request": "some request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "status_id": {
                                                "type": "integer"
                                            },
                                            "priority_id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "status_id": 1,
                                            "priority_id": 3,
                                            "title": "Some title",
                                            "description": "Some description"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "status_id": {
                                                    "type": "integer"
                                                },
                                                "priority_id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "status_id": 1,
                                                "priority_id": 3,
                                                "title": "Some title",
                                                "description": "Some description"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Access",
                "requestBody": {
                    "request": "some request",
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "email": "email@gmail.com",
                                            "password": "11111111"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL2F1dGgvbG9naW4iLCJpYXQiOjE3MDAwMDA0NzAsImV4cCI6MTcwMDAwNDA3MCwibmJmIjoxNzAwMDAwNDcwLCJqdGkiOiI4YkNIeGd5UUxhRlZ4czRLIiwic3ViIjoiMyIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.LWELjMOB8hLO5Qm_x0F-6d5ZcVn2cHLMel1hInWvOSE",
                                    "token_type": "bearer",
                                    "expires_id": 3600
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}